feat: add hash, color, and system utility modules
