+++ src/config.py
@@ -6,7 +6,7 @@ import shutil
 from pathlib import Path
 from typing import Any, Dict
 
-from .utils.helpers import log
+from .utils.system_utils import log
 
 
 class Config:
