+++ src/utils/__init__.py
@@ -2,23 +2,27 @@
 
 import os
 
-from .helpers import (
+from .system_utils import (
     log,
     open_path,
     slugify,
     strip_ansi,
+    get_system_info,
+    get_system_metrics,
+    run_with_spinner,
+    console,
+)
+from .hash_utils import (
     calc_data_hash,
     calc_hash,
     calc_hash_cached,
     calc_hashes,
-    get_system_info,
-    get_system_metrics,
+)
+from .color_utils import (
     lighten_color,
     darken_color,
     adjust_color,
     hex_brightness,
-    run_with_spinner,
-    console,
 )
 from .rainbow import RainbowBorder, NeonPulseBorder
 from .vm import launch_vm_debug
