+++ src/views/about_view.py
@@ -2,7 +2,7 @@
 About view - Application info
 """
 from ..components.widgets import info_label
-from ..utils import open_path
+from ..utils.system_utils import open_path
 from .base_view import BaseView
